Predatory_Journal_Detector/requirements.txt
# Predatory Journal Detector - Requirements
# Installation: pip install -r requirements.txt

# Web Scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
aiohttp>=3.9.0
fake-useragent>=1.4.0
requests-html>=0.10.0

# Machine Learning and Data Science
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.15.0

# Natural Language Processing
nltk>=3.8.0
textblob>=0.17.0
langdetect>=1.0.9
textstat>=0.7.3

# Web Framework and API
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10

# Database and Storage (optional for basic demo)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# pymongo>=4.5.0
# redis>=4.5.0

# Data Processing and Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
python-dateutil>=2.8.0
pytz>=2023.1
pydantic>=2.4.0

# Image Processing (optional)
Pillow>=10.0.0

# Network and Domain Analysis
whois>=0.9.27
dnspython>=2.4.0
python-whois>=0.8.0
tldextract>=5.1.0

# Model Interpretation (optional)
# shap>=0.42.0
# lime>=0.2.0.1

# Development and Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0

# Monitoring and Logging (optional)
# loguru>=0.7.0
# sentry-sdk>=1.35.0

# Configuration and Environment
pyyaml>=6.0.0
click>=8.1.0

# HTTP and URL utilities
validators>=0.22.0
furl>=2.1.3
urllib3>=2.1.0

# Async and Threading
# asyncio and concurrent.futures are built-in Python modules

# Scientific Computing (optional)
# sympy>=1.11.0
# statsmodels>=0.13.0

